Move write_introducer somewhere it can be used more widely
